+++ lib/compressedObject.js
@@ -1,10 +1,22 @@
 'use strict';
-function CompressedObject() {
-    this.compressedSize = 0;
-    this.uncompressedSize = 0;
-    this.crc32 = 0;
-    this.compressionMethod = null;
-    this.compressedContent = null;
+
+var utils = require('./utils');
+
+/**
+ * Represent a compressed object, with everything needed to decompress it.
+ * @constructor
+ * @param {number} compressedSize the size of the data compressed.
+ * @param {number} uncompressedSize the size of the data after decompression.
+ * @param {number} crc32 the crc32 of the decompressed file.
+ * @param {object} compression the type of compression, see lib/compressions.js.
+ * @param {String|ArrayBuffer|Uint8Array|Buffer} data the compressed data.
+ */
+function CompressedObject(compressedSize, uncompressedSize, crc32, compression, data) {
+    this.compressedSize = compressedSize;
+    this.uncompressedSize = uncompressedSize;
+    this.crc32 = crc32;
+    this.compression = compression;
+    this.compressedContent = data;
 }
 
 CompressedObject.prototype = {
@@ -14,7 +26,14 @@ CompressedObject.prototype = {
      * @return {Object} the decompressed content.
      */
     getContent: function() {
-        return null; // see implementation
+        var compressedFileData = utils.transformTo(this.compression.uncompressInputType, this.getCompressedContent());
+        var uncompressedFileData = this.compression.uncompress(compressedFileData);
+
+        if (uncompressedFileData.length !== this.uncompressedSize) {
+            throw new Error("Bug : uncompressed data size mismatch");
+        }
+
+        return uncompressedFileData;
     },
     /**
      * Return the compressed content in an unspecified format.
@@ -22,7 +41,22 @@ CompressedObject.prototype = {
      * @return {Object} the compressed content.
      */
     getCompressedContent: function() {
-        return null; // see implementation
+        return this.compressedContent;
+    },
+    /**
+     * Change the compression type (DEFLATE -> STORE for example).
+     * @param {Object} compression the new compression.
+     */
+    recompress: function (compression) {
+        if (this.compression.magic !== compression.magic) {
+            var content = this.getContent();
+            // need to decompress / recompress
+            this.compressedContent = compression.compress(utils.transformTo(compression.compressInputType, content));
+            this.compressedSize = this.compressedContent.length;
+            this.compression = compression;
+        }
     }
 };
 module.exports = CompressedObject;
+
+// vim: set shiftwidth=4 softtabstop=4:
